requirements.txt
Flask
Flask-Cors
requests
mysql-connector-python